refactor: move deploy checks
